+++ src/internal/privateTypes.ts
@@ -17,6 +17,8 @@ interface NativeConstants {
   deviceId: string;
   deviceType: DeviceType;
   isTablet: boolean;
+  isLowRamDevice: boolean;
+  isDisplayZoomed: boolean;
   model: string;
   systemName: string;
   systemVersion: string;
@@ -75,6 +77,8 @@ interface ExposedNativeMethods {
   getHardwareSync: () => string;
   getHost: () => Promise<string>;
   getHostSync: () => string;
+  getHostNames: () => Promise<string[]>;
+  getHostNamesSync: () => string[];
   getIncremental: () => Promise<string>;
   getIncrementalSync: () => string;
   getInstallerPackageName: () => Promise<string>;
@@ -91,8 +95,6 @@ interface ExposedNativeMethods {
   getMacAddressSync: () => string;
   getMaxMemory: () => Promise<number>;
   getMaxMemorySync: () => number;
-  getPhoneNumber: () => Promise<string>;
-  getPhoneNumberSync: () => string;
   getPreviewSdkInt: () => Promise<number>;
   getPreviewSdkIntSync: () => number;
   getProduct: () => Promise<string>;
@@ -137,6 +139,10 @@ interface ExposedNativeMethods {
   isEmulatorSync: () => boolean;
   isHeadphonesConnected: () => Promise<boolean>;
   isHeadphonesConnectedSync: () => boolean;
+  isWiredHeadphonesConnected: () => Promise<boolean>;
+  isWiredHeadphonesConnectedSync: () => boolean;
+  isBluetoothHeadphonesConnected: () => Promise<boolean>;
+  isBluetoothHeadphonesConnectedSync: () => boolean;
   isLocationEnabled: () => Promise<boolean>;
   isLocationEnabledSync: () => boolean;
   isPinOrFingerprintSet: () => Promise<boolean>;
@@ -147,6 +153,8 @@ interface ExposedNativeMethods {
   isKeyboardConnectedSync: () => boolean;
   isTabletMode: () => Promise<boolean>;
   syncUniqueId: () => Promise<string>;
+  getSupportedMediaTypeList: () => Promise<string[]>;
+  getSupportedMediaTypeListSync: () => string[];
 }
 
 export interface DeviceInfoNativeModule
@@ -179,6 +187,8 @@ export interface DeviceInfoModule extends ExposedNativeMethods {
   isLandscape: () => Promise<boolean>;
   isLandscapeSync: () => boolean;
   isTablet: () => boolean;
+  isLowRamDevice: () => boolean;
+  isDisplayZoomed: () => boolean;
   supported32BitAbis: () => Promise<string[]>;
   supported32BitAbisSync: () => string[];
   supported64BitAbis: () => Promise<string[]>;
@@ -194,6 +204,8 @@ export interface DeviceInfoModule extends ExposedNativeMethods {
   usePowerState: () => Partial<PowerState>;
   useManufacturer: () => AsyncHookResult<string>;
   useIsHeadphonesConnected: () => AsyncHookResult<boolean>;
+  useIsWiredHeadphonesConnected: () => AsyncHookResult<boolean>;
+  useIsBluetoothHeadphonesConnected: () => AsyncHookResult<boolean>;
   useBrightness: () => number | null;
 }